requirements.txt
flask
google-genai
pillow
python-dotenv
waitress
flask-cors
gunicorn